api/requirements.txt
python-dotenv
flask
requests
# authlib